Move Aztec's `TextViewAttachmentDelegate` to a DI'ed property.
